minigo-project/requirements.txt
antlr4-python3-runtime
pytest
pytest-antlr